Integrate real-time MT5 feed into main dashboard
