+++ requirements.txt
@@ -36,3 +36,4 @@ urllib3==2.2.3
 vine==5.1.0
 wcwidth==0.2.13
 whitenoise==6.7.0
+PyYAML==6.0.2
